+++ src/devServer.js
@@ -1,9 +1,10 @@
+import opn from 'opn'
 import webpack from 'webpack'
 import WebpackDevServer from 'webpack-dev-server'
 import merge from 'webpack-merge'
 
 import debug from './debug'
-import {deepToString} from './utils'
+import {deepToString, typeOf} from './utils'
 
 /**
  * Use Webpack Dev Server to build and serve assets using Webpack's watch mode,
@@ -11,10 +12,10 @@ import {deepToString} from './utils'
  *
  * Static content is handled by CopyPlugin.
  */
-export default function devServer(webpackConfig, serverConfig, cb) {
+export default function devServer(webpackConfig, serverConfig, url, cb) {
   let compiler = webpack(webpackConfig)
 
-  let {host, port, ...otherServerConfig} = serverConfig
+  let {host, open, port, ...otherServerConfig} = serverConfig
 
   let webpackDevServerOptions = merge({
     headers: {
@@ -34,16 +35,13 @@ export default function devServer(webpackConfig, serverConfig, cb) {
   debug('webpack dev server options: %s', deepToString(webpackDevServerOptions))
 
   let server = new WebpackDevServer(compiler, webpackDevServerOptions)
-
-  function onServerStart(err) {
-    if (err) cb(err)
-  }
-
-  // Only provide host config if it was explicitly specified by the user
-  if (host) {
-    server.listen(port, host, onServerStart)
-  }
-  else {
-    server.listen(port, onServerStart)
-  }
+  server.listen(port, host, (err) => {
+    if (err) return cb(err)
+    if (open) {
+      // --open
+      if (typeOf(open) === 'boolean') opn(url)
+      // --open=firefox
+      else opn(url, {app: open})
+    }
+  })
 }